backend/requirements.txt
Flask
Flask-CORS
pymongo
python-dotenv
requests
python-ffmpeg
flask-limiter
marshmallow
flask-marshmallow
marshmallow-sqlalchemy 